Moved all diff collectors to their package. Easier to find them in the IDE :)
